@@ -1,8 +1,11 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using MyCourse.Models.Exceptions;
+using MyCourse.Models.InputModels;
 using MyCourse.Models.Services.Application;
 using MyCourse.Models.ViewModels;
+
 
 namespace MyCourse.Controllers
 {
@@ -12,23 +15,88 @@ namespace MyCourse.Controllers
         public CoursesController(ICachedCourseService courseService)
         {
             this.courseService = courseService;
-
         }
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(CourseListInputModel input)
         {
             ViewData["Title"] = "Catalogo dei corsi";
+            ListViewModel<CourseViewModel> courses = await courseService.GetCoursesAsync(input);
+            
+            CourseListViewModel viewModel = new CourseListViewModel
+            {
+                Courses = courses,
+                Input = input
+            };
 
-            List<CourseViewModel> courses = await courseService.GetCoursesAsync();
-            return View(courses);
+            return View(viewModel);
         }
 
         public async Task<IActionResult> Detail(int id)
         {
-
             CourseDetailViewModel viewModel = await courseService.GetCourseAsync(id);
             ViewData["Title"] = viewModel.Title;
             return View(viewModel);
         }
 
+        public IActionResult Create()
+        {
+           ViewData["Title"] = "Nuovo corso";
+           var inputModel = new CourseCreateinputModel();
+            return View(inputModel);
+        }
+        [HttpPost]
+        public async Task<IActionResult> Create(CourseCreateinputModel inputModel)
+        {
+           if(ModelState.IsValid)
+           {
+               try
+               {
+                   CourseDetailViewModel course = await courseService.CreateCourseAsync(inputModel);
+                   TempData["ConfirmationMessage"] = "Ok! Il tuo corso è stato creato, ore perchè non inserisci anche gli altri dati?";
+                   return RedirectToAction(nameof(Edit),new{id = course.Id});
+               }
+               catch (CourseTitleUnavailableException)
+               {
+                  ModelState.AddModelError(nameof(CourseDetailViewModel.Title),"Questo titolo già esiste");
+               }
+           }
+               ViewData["Title"] = "Nuovo corso";
+               return View(inputModel);
+
+           
+            //coinvolgere un servizio applicativo in modo che il corso venga creato
+        }
+        public async Task<IActionResult> IsTitleAvailable(string title, int id=0)
+        {
+            bool result = await courseService.IsTitleAvailableAsync(title,id);
+            return Json(result);
+        }
+
+        public async Task<IActionResult> Edit(int id)
+        {
+            ViewData["Title"] = "Modifica corso";
+            CourseEditInputModel inputModel = await courseService.GetCourseForEditingAsync(id);
+               return View(inputModel);
+        }
+        
+        [HttpPost]
+        public async Task<IActionResult> Edit(CourseEditInputModel inputModel)
+        {
+            if(ModelState.IsValid)
+            {
+                try
+                {
+                  CourseDetailViewModel course = await courseService.EditCourseAsync(inputModel);
+                  TempData["ConfirmationMessage"] = "I dati sono stati salvati con successo";
+                  return RedirectToAction(nameof(Detail), new{id = inputModel.Id});
+                }
+                catch (CourseTitleUnavailableException)
+                {
+                    ModelState.AddModelError(nameof(CourseDetailViewModel.Title),"Questo titolo già esiste");
+                }
+
+            }
+            ViewData["Title"] = "Modifica corso";
+               return View(inputModel);
+        }
     }
 }
