@@ -1,6 +1,7 @@
 using System;
 using Microsoft.AspNetCore.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
+using MyCourse.Models.Exceptions;
 
 namespace MyCourse.Controllers
 {
@@ -8,9 +9,18 @@ namespace MyCourse.Controllers
     {
         public IActionResult Index()
         {
-            
+            var feature = HttpContext.Features.Get<IExceptionHandlerPathFeature>();
+            switch(feature.Error)
+            {
+                case CourseNotFoundException exc:
+                    ViewData["Title"] = "Corso non trovato";
+                    Response.StatusCode = 404;
+                    return View("CourseNotFound");
+
+                default:
                     ViewData["Title"] = "Errore";
                     return View();
             }
         }
     }
+}
