@@ -1,13 +1,27 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using MyCourse.Models.InputModels;
+using MyCourse.Models.Services.Application;
+using MyCourse.Models.ViewModels;
 
 namespace MyCourse.Controllers
 {
-    public class HomeController:Controller
+    public class HomeController : Controller
     {
-        public IActionResult Index()
+        public async Task<IActionResult> Index([FromServices] ICachedCourseService courseService)
         {
-            //return Content("Ciao");
-            return View();
+            ViewData["Title"] = "Benvenuto su MyCourse!";
+            List<CourseViewModel> bestRatingCourses = await courseService.GetBestRatingCoursesAsync();
+            List<CourseViewModel> mostRecentCourses = await courseService.GetMostRecentCoursesAsync();
+
+            HomeViewModel viewModel = new HomeViewModel
+            {
+                BestRatingCourses = bestRatingCourses,
+                MostRecentCourses = mostRecentCourses
+            };
+
+            return View(viewModel);
         }
     }
 }
