@@ -1,5 +1,6 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
+using MyCourse.Models.Enums;
 using MyCourse.Models.ValueTypes;
 
 namespace MyCourse.Models.Entities
@@ -15,11 +16,14 @@ namespace MyCourse.Models.Entities
 			if (string.IsNullOrWhiteSpace(author))
             {
                 throw new ArgumentException("The course must have an author");
-            }   
+            }
 
             Title = title;
             Author = author;
             Lessons = new HashSet<Lesson>();
+            CurrentPrice = new Money(Currency.EUR, 0);
+            FullPrice = new Money(Currency.EUR, 0);
+            ImagePath = "/Courses/default.png";
         }
 
         public int Id { get; private set; }
@@ -41,22 +45,47 @@ namespace MyCourse.Models.Entities
             Title = newTitle;
         }
 
-        public void ChangePrices(Money newFullPrice, Money newDiscountPrice)
+        public void ChangePrices(Money newFullPrice, Money newCurrentPrice)
         {
-            if (newFullPrice == null || newDiscountPrice == null)
+            if (newFullPrice == null || newCurrentPrice == null)
             {
                 throw new ArgumentException("Prices can't be null");
             }
-            if (newFullPrice.Currency != newDiscountPrice.Currency)
+            if (newFullPrice.Currency != newCurrentPrice.Currency)
             {
                 throw new ArgumentException("Currencies don't match");
             }
-            if (newFullPrice.Amount < newDiscountPrice.Amount)
+            if (newFullPrice.Amount < newCurrentPrice.Amount)
             {
                 throw new ArgumentException("Full price can't be less than the current price");
             }
             FullPrice = newFullPrice;
-            CurrentPrice = newDiscountPrice;
+            CurrentPrice = newCurrentPrice;
+        }
+
+        public void ChangeEmail(string newEmail)
+        {
+            if (string.IsNullOrEmpty(newEmail))
+            {
+                throw new ArgumentException("Email can't be empty");
+            }
+            Email = newEmail;
+        }
+
+        public void ChangeDescription(string newDescription)
+        {
+            if (newDescription != null)
+            {
+                if (newDescription.Length < 20)
+                {
+                    throw new Exception("Description is too short");
+                }
+                else if (newDescription.Length > 4000)
+                {
+                    throw new Exception("Description is too long");
+                }
+            }
+            Description = newDescription;
         }
 
         public virtual ICollection<Lesson> Lessons { get; private set; }
