@@ -2,6 +2,6 @@ namespace MyCourse.Models.Options
 {
     public class ConnectionStringsOptions
     {
-        public string Default {get; set;}
+        public string Default { get; set; }
     }
 }
