@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using MyCourse.Controllers;
 using MyCourse.Models.Enums;
 using MyCourse.Models.ValueTypes;
 using MyCourse.Models.ViewModels;
@@ -13,7 +12,7 @@ namespace MyCourse.Models.Services.Application
         {
             var courseList = new List<CourseViewModel>();
             var rand = new Random();
-            for(int i= 1; i <= 20; i++)
+            for (int i = 1; i <= 20; i++)
             {
                 var price = Convert.ToDecimal(rand.NextDouble() * 10 + 10);
                 var course = new CourseViewModel
@@ -21,9 +20,9 @@ namespace MyCourse.Models.Services.Application
                     Id = i,
                     Title = $"Corso {i}",
                     CurrentPrice = new Money(Currency.EUR, price),
-                    FullPrice = new Money(Currency.EUR, rand.NextDouble()>0.5 ? price:price-1),
+                    FullPrice = new Money(Currency.EUR, rand.NextDouble() > 0.5 ? price : price - 1),
                     Author = "Nome cognome",
-                    Rating = rand.NextDouble() * 5.0,
+                    Rating = rand.Next(10, 50) / 10.0,
                     ImagePath = "/logo.svg"
                 };
                 courseList.Add(course);
@@ -33,29 +32,30 @@ namespace MyCourse.Models.Services.Application
 
         public CourseDetailViewModel GetCourse(int id)
         {
-                var rand = new Random();
-                var price = Convert.ToDecimal(rand.NextDouble() * 10 + 10);
-                var course = new CourseDetailViewModel
-                {
-                    Id = id,
-                    Title = $"Corso {id}",
-                    CurrentPrice = new Money(Currency.EUR, price),
-                    FullPrice = new Money(Currency.EUR, rand.NextDouble()>0.5 ? price:price-1),
-                    Author = "Nome cognome",
-                    Rating = rand.Next(10, 50) / 10.0,
-                    ImagePath = "/logo.svg",
-                    Description = $"Descrizione {id}",
-                    Lessons = new List<LessonViewModel>()
-                };
+            var rand = new Random();
+            var price = Convert.ToDecimal(rand.NextDouble() * 10 + 10);
+            var course = new CourseDetailViewModel
+            {
+                Id = id,
+                Title = $"Corso {id}",
+                CurrentPrice = new Money(Currency.EUR, price),
+                FullPrice = new Money(Currency.EUR, rand.NextDouble() > 0.5 ? price : price - 1),
+                Author = "Nome cognome",
+                Rating = rand.Next(10, 50) / 10.0,
+                ImagePath = "/logo.svg",
+                Description = $"Descrizione {id}",
+                Lessons = new List<LessonViewModel>()
+            };
 
-                for (var i=1; i<=5; i++){
-                    var lesson = new LessonViewModel {
-                        Title = $"Lezione {i}",
-                        Duration = TimeSpan.FromSeconds(rand.Next(40,90))
-                    };
-                    course.Lessons.Add(lesson);
-                }
-                return course;
+            for (var i = 1; i <= 5; i++) {
+                var lesson = new LessonViewModel {
+                    Title = $"Lezione {i}",
+                    Duration = TimeSpan.FromSeconds(rand.Next(40, 90))
+                };
+                course.Lessons.Add(lesson);
+            }
+
+            return course;
         }
     }
 }
