@@ -1,7 +1,15 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using MyCourse.Models.Entities;
+using MyCourse.Models.Exceptions;
+using MyCourse.Models.InputModels;
+using MyCourse.Models.Options;
 using MyCourse.Models.Services.Infrastructure;
 using MyCourse.Models.ViewModels;
 
@@ -9,10 +17,14 @@ namespace MyCourse.Models.Services.Application
 {
     public class EfCoreCourseService : ICourseService
     {
+        private readonly ILogger<EfCoreCourseService> logger;
         private readonly MyCourseDbContext dbContext;
+        private readonly IOptionsMonitor<CoursesOptions> coursesOptions;
 
-        public EfCoreCourseService(MyCourseDbContext dbContext)
+        public EfCoreCourseService(ILogger<EfCoreCourseService> logger, MyCourseDbContext dbContext, IOptionsMonitor<CoursesOptions> coursesOptions)
         {
+            this.coursesOptions = coursesOptions;
+            this.logger = logger;
             this.dbContext = dbContext;
         }
 
@@ -23,24 +35,134 @@ namespace MyCourse.Models.Services.Application
                 .Include(course => course.Lessons)
                 .Where(course => course.Id == id)
                 .Select(course => CourseDetailViewModel.FromEntity(course)); //Usando metodi statici come FromEntity, la query potrebbe essere inefficiente. Mantenere il mapping nella lambda oppure usare un extension method personalizzato
-            
-            CourseDetailViewModel viewModel = await queryLinq.SingleAsync();
-                                                           //.FirstOrDefaultAsync(); //Restituisce null se l'elenco è vuoto e non solleva mai un'eccezione
-                                                           //.SingleOrDefaultAsync(); //Tollera il fatto che l'elenco sia vuoto e in quel caso restituisce null, oppure se l'elenco contiene più di 1 elemento, solleva un'eccezione
-                                                           //.FirstAsync(); //Restituisce il primo elemento, ma se l'elenco è vuoto solleva un'eccezione
-                
+
+            CourseDetailViewModel viewModel = await queryLinq.FirstOrDefaultAsync();
+            //.FirstOrDefaultAsync(); //Restituisce null se l'elenco è vuoto e non solleva mai un'eccezione
+            //.SingleOrDefaultAsync(); //Tollera il fatto che l'elenco sia vuoto e in quel caso restituisce null, oppure se l'elenco contiene più di 1 elemento, solleva un'eccezione
+            //.FirstAsync(); //Restituisce il primo elemento, ma se l'elenco è vuoto solleva un'eccezione
+            //.SingleAsync(); //Restituisce il primo elemento, ma se l'elenco è vuoto o contiene più di un elemento, solleva un'eccezione
+
+            if (viewModel == null)
+            {
+                logger.LogWarning("Course {id} not found", id);
+                throw new CourseNotFoundException(id);
+            }
+
             return viewModel;
         }
 
-        public async Task<List<CourseViewModel>> GetCoursesAsync()
+        public async Task<List<CourseViewModel>> GetBestRatingCoursesAsync()
         {
-            IQueryable<CourseViewModel> queryLinq = dbContext.Courses
-                .AsNoTracking()
-                .Select(course => CourseViewModel.FromEntity(course)); //Usando metodi statici come FromEntity, la query potrebbe essere inefficiente. Mantenere il mapping nella lambda oppure usare un extension method personalizzato
+            CourseListInputModel inputModel = new CourseListInputModel(
+                search: "",
+                page: 1,
+                orderby: "Rating",
+                ascending: false,
+                limit: coursesOptions.CurrentValue.InHome,
+                orderOptions: coursesOptions.CurrentValue.Order);
 
-            List<CourseViewModel> courses = await queryLinq.ToListAsync(); //La query al database viene inviata qui, quando manifestiamo l'intenzione di voler leggere i risultati
+                ListViewModel<CourseViewModel> result = await GetCoursesAsync(inputModel);
+                return result.Results;
+        }
+        
+        public async Task<List<CourseViewModel>> GetMostRecentCoursesAsync()
+        {
+            CourseListInputModel inputModel = new CourseListInputModel(
+                search: "",
+                page: 1,
+                orderby: "Id",
+                ascending: false,
+                limit: coursesOptions.CurrentValue.InHome,
+                orderOptions: coursesOptions.CurrentValue.Order);
 
-            return courses;
+                ListViewModel<CourseViewModel> result = await GetCoursesAsync(inputModel);
+                return result.Results;
+        }
+
+        public async Task<ListViewModel<CourseViewModel>> GetCoursesAsync(CourseListInputModel model)
+        {
+            IQueryable<Course> baseQuery = dbContext.Courses;
+
+            baseQuery = (model.OrderBy, model.Ascending) switch
+            {
+                ("Title", true) => baseQuery.OrderBy(course => course.Title),
+                ("Title", false) => baseQuery.OrderByDescending(course => course.Title),               
+                ("Rating", true) => baseQuery.OrderBy(course => course.Rating),
+                ("Rating", false) => baseQuery.OrderByDescending(course => course.Rating),
+                ("CurrentPrice", true) => baseQuery.OrderBy(course => course.CurrentPrice.Amount),
+                ("CurrentPrice", false) => baseQuery.OrderByDescending(course => course.CurrentPrice.Amount),
+                ("Id", true) => baseQuery.OrderBy(course => course.Id),
+                ("Id", false) => baseQuery.OrderByDescending(course => course.Id),
+                _ => baseQuery
+            };
+
+            IQueryable<Course> queryLinq = baseQuery
+                .Where(course => course.Title.Contains(model.Search))
+                .AsNoTracking();
+
+            List<CourseViewModel> courses = await queryLinq                
+                .Skip(model.Offset)
+                .Take(model.Limit)
+                .Select(course => CourseViewModel.FromEntity(course)) //Usando metodi statici come FromEntity, la query potrebbe essere inefficiente. Mantenere il mapping nella lambda oppure usare un extension method personalizzato
+                .ToListAsync(); //La query al database viene inviata qui, quando manifestiamo l'intenzione di voler leggere i risultati
+
+            int totalCount = await queryLinq.CountAsync();
+
+            ListViewModel<CourseViewModel> result = new ListViewModel<CourseViewModel>
+            {
+                Results = courses,
+                TotalCount = totalCount
+            };
+
+            return result;
+        }
+
+
+        public async Task<CourseDetailViewModel> CreateCourseAsync(CourseCreateinputModel inputModel)
+        {
+            string title = inputModel.Title;
+            string author = "Mario Rossi";
+
+            var course = new Course(title, author);
+            dbContext.Add(course);
+            await dbContext.SaveChangesAsync();
+            
+
+            return CourseDetailViewModel.FromEntity(course);
+        }
+        public async Task<bool> IsTitleAvailableAsync(string title,int id)
+        {
+            bool titleExists = await dbContext.Courses.AnyAsync(course => EF.Functions.Like(course.Title, title));
+            return !titleExists;
+        }
+
+        public Task<CourseEditInputModel> GetCourseForEditingAsync(int id)
+        {
+            throw new NotImplementedException();
+        }
+
+        public async Task<CourseDetailViewModel> EditCourseAsync(CourseEditInputModel inputModel)
+        {
+            Course course = await dbContext.Courses.FindAsync(inputModel.Id);
+
+            course.ChangeTitle(inputModel.Title);
+            course.ChangePrices(inputModel.FullPrice, inputModel.CurrentPrice);
+            course.ChangeDescription(inputModel.Description);
+            course.ChangeEmail(inputModel.Email);
+            
+           /* string imagePath = await imagePersiser.SaveCourseImageAsync(inputModel.Id, inputModel.Image);
+            course.ChangeImagePath(imagePath;) */
+            //dbContext.Update(course);
+
+            try 
+            {
+              await dbContext.SaveChangesAsync();
+            }
+            catch(DbUpdateException exc) when ((exc.InnerException as SqliteException)?.SqliteErrorCode == 19)
+            {
+              throw new CourseTitleUnavailableException(inputModel.Title, exc);
+            }
+            return CourseDetailViewModel.FromEntity(course);
         }
     }
 }
