@@ -1,7 +1,10 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using MyCourse.Models.ViewModels;
+
 namespace MyCourse.Models.Services.Application
 {
     public interface ICachedCourseService : ICourseService
     {
-         
     }
 }
