@@ -1,12 +1,19 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using MyCourse.Models.InputModels;
 using MyCourse.Models.ViewModels;
 
 namespace MyCourse.Models.Services.Application
 {
     public interface ICourseService
     {
-         Task<List<CourseViewModel>> GetCoursesAsync();
-         Task<CourseDetailViewModel> GetCourseAsync(int id);
+        Task<ListViewModel<CourseViewModel>> GetCoursesAsync(CourseListInputModel model);
+        Task<CourseDetailViewModel> GetCourseAsync(int id);
+        Task<List<CourseViewModel>> GetMostRecentCoursesAsync();
+        Task<List<CourseViewModel>> GetBestRatingCoursesAsync();
+        Task<CourseEditInputModel> GetCourseForEditingAsync(int id);
+        Task<CourseDetailViewModel> CreateCourseAsync(CourseCreateinputModel inputModel);
+        Task<CourseDetailViewModel> EditCourseAsync(CourseEditInputModel inputModel);
+        Task<bool> IsTitleAvailableAsync(string title, int excludeId);
     }
 }
