@@ -7,7 +7,6 @@ namespace MyCourse.Models.Services.Infrastructure
 {
     public partial class MyCourseDbContext : DbContext
     {
-        
 
         public MyCourseDbContext(DbContextOptions<MyCourseDbContext> options)
             : base(options)
@@ -17,36 +16,38 @@ namespace MyCourse.Models.Services.Infrastructure
         public virtual DbSet<Course> Courses { get; set; }
         public virtual DbSet<Lesson> Lessons { get; set; }
 
-        
-
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
-            modelBuilder.HasAnnotation("ProductVersion", "2.2.6-servicing-10079");
+            modelBuilder.HasAnnotation("ProductVersion", "2.2.4-servicing-10062");
 
             modelBuilder.Entity<Course>(entity =>
             {
-                entity.ToTable("Courses"); //Superfluo se la tabella si chiama come la proprietà
-                entity.HasKey(course => course.Id);
+                entity.ToTable("Courses"); //Superfluo se la tabella si chiama come la proprietà che espone il DbSet
+                entity.HasKey(course => course.Id); //Superfluo se la proprietà si chiama Id oppure CoursesId
+                //entity.HasKey(course => new { course.Id, course.Author }); //Per chiavi primarie composite (è importante rispettare l'ordine dei campi)
 
-//mapping per gli owned types
+                //Mapping per gli owned types
                 entity.OwnsOne(course => course.CurrentPrice, builder => {
                     builder.Property(money => money.Currency)
                     .HasConversion<string>()
-                    .HasColumnName("CurrentPrice_Currency");
-                    builder.Property(money => money.Amount).HasColumnName("CurrentPrice_Amount");
-                });
-                entity.OwnsOne(course => course.FullPrice, builder => {
-                    builder.Property(money => money.Currency).HasConversion<string>();
-                    
+                    .HasColumnName("CurrentPrice_Currency"); //Superfluo perché le nostre colonne seguono già la convenzione di nomi
+                    builder.Property(money => money.Amount).HasColumnName("CurrentPrice_Amount") //Superfluo perché le nostre colonne seguono già la convenzione di nomi
+                    .HasConversion<float>()
+                    .HasColumnName("CurrentPrice_Amount");
                 });
 
-//mapping per le relazioni
-               entity.HasMany(course => course.Lessons)
-                     .WithOne(lesson => lesson.Course)
-                     .HasForeignKey(lesson => lesson.CourseId); //superflua se la proprieta si chiama CourseId
-                
+                entity.OwnsOne(course => course.FullPrice, builder => {
+                    builder.Property(money => money.Currency).HasConversion<string>();
+                });
+
+                //Mapping per le relazioni
+                entity.HasMany(course => course.Lessons)
+                      .WithOne(lesson => lesson.Course)
+                      .HasForeignKey(lesson => lesson.CourseId); //Superflua se la proprietà si chiama CourseId
+
                 #region Mapping generato automaticamente dal tool di reverse engineering
-                /* entity.Property(e => e.Id).ValueGeneratedNever();
+                /*
+                entity.Property(e => e.Id).ValueGeneratedNever();
 
                 entity.Property(e => e.Author)
                     .IsRequired()
@@ -85,15 +86,14 @@ namespace MyCourse.Models.Services.Infrastructure
                 entity.Property(e => e.Title)
                     .IsRequired()
                     .HasColumnType("TEXT (100)");
-                     */
-                    #endregion
+                    */
+                #endregion
             });
 
             modelBuilder.Entity<Lesson>(entity =>
             {
+                //Nessun mapping necessario qui, perché stiamo rispettando le convenzioni di nomi
 
-               entity.HasOne(lesson => lesson.Course)
-                     .WithMany(course => course.Lessons);
 
                 #region Mapping generato automaticamente dal tool di reverse engineering
                 /*
@@ -113,10 +113,9 @@ namespace MyCourse.Models.Services.Infrastructure
                 entity.HasOne(d => d.Course)
                     .WithMany(p => p.Lessons)
                     .HasForeignKey(d => d.CourseId);
-            
                 */
-            #endregion
-            }); 
+                #endregion
+            });
         }
     }
 }
