@@ -48,7 +48,7 @@ namespace MyCourse.Models.ValueTypes
         
         public override string ToString()
         {
-            return $"{Currency} {Amount:#.00}";
+            return $"{Currency} {Amount:0.00}";
         }
     }
 }
