@@ -2,41 +2,24 @@
 <section class="course-info">
     <div class="row">
         <div class="col-md-9">
-            <h1>@Model.Title</h1>
+            <h1>@Model.Title <a class="btn btn-outline-primary btn-sm" asp-action="Edit" asp-route-id="@Model.Id"><i class="fa fa-pencil-alt"></i>Modifica</a></h1>
         </div>
         <div class="col-md-3 d-flex align-items-end">
             <a href="#" class="btn btn-warning btn-lg btn-block subscribe-course">Iscriviti</a>
         </div>
     </div>
     <div class="row">
-        <div class="col-md-9">
-            <p>@Model.Description</p>
+        <div class="col-md-9" html-sanitize>
+            @Model.Description
         </div>
         <div class="col-md-3">
             <aside>
                 <author>di @Model.Author</author>
                 <div class="rating" title="@Model.Rating">
-                    @for(int i = 1; i <= 5; i++)
-                    {
-                        if (Model.Rating >= i)
-                        {
-                            <i class="fas fa-star"></i>
-                        }
-                        else if (Model.Rating > i - 1)
-                        {
-                            <i class="fas fa-star-half-alt"></i>
-                        }
-                        else
-                        {
-                            <i class="far fa-star"></i>
-                        }
-                    }
+                    <rating value="@Model.Rating"></rating>
                 </div>
                 <div class="price">
-                    @Model.CurrentPrice
-                    @if(!Model.CurrentPrice.Equals(Model.FullPrice)) {
-                        <br><s>@Model.FullPrice</s>
-                    }
+                    <price current-price="@Model.CurrentPrice" full-price="@Model.FullPrice"></price>
                 </div>
                 <img src="@Model.ImagePath" class="img-fluid course-image" alt="@Model.Title">
             </aside>
